Use intptr_t instead of int to cast from char* for Snow Leopard.
git-svn-id: https://google-glog.googlecode.com/svn/trunk@72 eb4d4688-79bd-11dd-afb4-1d65580434c0

+++ src/logging.cc
@@ -1755,7 +1755,7 @@ int posix_strerror_r(int err, char *buf, size_t len) {
     } else {
       buf[0] = '\000';
 #if defined(OS_MACOSX) || defined(OS_FREEBSD) || defined(OS_OPENBSD)
-      if (reinterpret_cast<int>(rc) < sys_nerr) {
+      if (reinterpret_cast<intptr_t>(rc) < sys_nerr) {
         // This means an error on MacOSX or FreeBSD.
         return -1;
       }
